feat: airtableBaseDao methods are now fully async
BREAKING CHANGE:
airtableBaseDao methods are now fully async.
This is to prevent a possible bug when airtableBaseDao is not properly
initialized, so it would default to JSON connector and point to stale json
files.
Now it requires `primaryConnector` configuration, which is allowed to be an async
connector.
Previously, only a sync connector was allowed to be primary, and was
confusingly called `lazyConnector`.